+++ src/OPI/internal/opi_pluginprocs.h
@@ -45,10 +45,10 @@ namespace OPI
 	// cpp propagator interface function
 	typedef Propagator* (*pluginPropagatorFunction)(OPI_Host host);
 	// c interface propagation function
-	typedef ErrorCode (*pluginPropagateFunction)(OPI_Propagator propagator, OPI_Population data, double julian_day, float dt);
+    typedef ErrorCode (*pluginPropagateFunction)(OPI_Propagator propagator, OPI_Population data, double julian_day, double dt);
 
 	// c interface propagation function
-	typedef ErrorCode (*pluginPropagateFunctionIndexed)(OPI_Propagator propagator, OPI_Population data, int* indices, int index_size, double julian_day, float dt);
+    typedef ErrorCode (*pluginPropagateFunctionIndexed)(OPI_Propagator propagator, OPI_Population data, int* indices, int index_size, double julian_day, double dt);
 
 	// cpp distance query interface function
 	typedef DistanceQuery* (*pluginDistanceQueryFunction)(OPI_Host host);
