+++ src/OPI/internal/opi_propagator_plugin.h
@@ -35,8 +35,8 @@ namespace OPI
 
 			virtual ErrorCode enable();
 			virtual ErrorCode disable();
-			virtual ErrorCode runPropagation(Population& data, double julian_day, float dt );
-			virtual ErrorCode runIndexedPropagation(Population& data, int* indices, int index_size, double julian_day, float dt );
+            virtual ErrorCode runPropagation(Population& data, double julian_day, double dt);
+            virtual ErrorCode runIndexedPropagation(Population& data, int* indices, int index_size, double julian_day, double dt);
 			virtual int requiresCUDA();
 		private:
 			Plugin* plugin;
